Include tests (Rspec) for existing functionality
We have included 23 tests, all passing, for existing functionality of
Gilded Rose's update_quality method. The original test has been removed.
A spec_helper patches the GildedRose.initialize method, so we can inject
a custom Item object for test into the @items parameter.

Normal, Brie, Legendary, Backstage Passes are all tested for a drop in
sell in date, not exceeding 50 quality on increases, not going past 0
quality in decreases, and boundaries of being on the sell_in and past
the sell_in date.

Aged Brie increases in quality twice as fast when its past its sell in
date. This is not clear in the original specification so the test case
matches the code.

Changes to be committed:
  modified:   gilded_rose_spec.rb
  new file:   gilded_rose_spec_helper.rb