Refactor common conditionals and move methods
Incrementing quality has been moved to a separate method and includes
the if quality < 50 conditional. Decrementing quality has been moved to
a separate method with the if quality > 0 conditional.

Redundant @items = [] has been removed. The for loop for @items has been
changed to an each do iterator. Syntax has been changed to reflect this.

Legendary items are caught at the beginning of the and a return is
called.

Decrementing sell in has been moved to its own method. Although may not
actually reduce complexity any more.