Alternative refactor to GildedRose
This is alternative to the branch Third Refactor which doesn't extend
the Item class. This could be preferable because the Item class is
deemed untouchable in the spec.

This commit creates an Update class which whose method update finds the
right class for updating each item. This also has classes for the types
of item

* ItemUpdater - standard updater
  * NormalUpdater - as Item with -1 quality increments
  * ImprovingUpdater - as Item with +1 quality increments
    * TicketsUpdater - as Improving, with business logic for tickets
* LegendaryUpdater - not subclassing Item - no change applied to Item